support up ciot(distinguish up/cp proc by func use_cp_ciot)

+++ build.sh
@@ -4,6 +4,7 @@ PROJECT_ROOT_DIR=$(dirname `readlink -f $0`)
 echo $PROJECT_ROOT_DIR
 
 CROSSCOMPILE=0
+REBUILD=0
 
 while [[ $# -gt 0 ]]
 do
@@ -14,12 +15,16 @@ do
             shift
 	    ;;
 
+        -r)
+            REBUILD=1
+            shift
+	    ;;
+
     	*)
             echo 'ERROR: not support options'
             exit -1
     esac
 done
-
 
 if [ $CROSSCOMPILE -eq 0 ]; then
     MESON_CROSSCOMPILE=""
@@ -38,32 +43,35 @@ else
     LIBUSB_CROSSCOMPILE_OPTION="--host=aarch64-linux-gnu CC=aarch64-linux-gnu-gcc CXX=aarch64-linux-gnu-g++ AR=aarch64-linux-gnu-ar RANLIB=aarch64-linux-gnu-ranlib STRIP=aarch64-linux-gnu-strip"
 fi
 
-THIRD_PARTY_DIR=$PROJECT_ROOT_DIR/third_party
-
 # libmbedtls start
-
+THIRD_PARTY_DIR=$PROJECT_ROOT_DIR/third_party
 libmbedtls_src_dir=$THIRD_PARTY_DIR/mbedtls
 libmbedtls_build_dir=$libmbedtls_src_dir/build
-rm -rf $libmbedtls_build_dir
 libmbedtls_install_dir="$THIRD_PARTY_DIR/mbedtls-install/"
-rm -rf $libmbedtls_install_dir
+if [ $REBUILD -eq 1 ] || [ ! -d $libmbedtls_install_dir ]; then
+  if [ -d $libmbedtls_build_dir ]; then
+	rm -rf $libmbedtls_build_dir
+  fi
+  if [ -d $libmbedtls_install_dir ]; then
+    rm -rf $libmbedtls_install_dir
+  fi
 
-echo "build libmbedtls"
-build_type="Release"
+  echo "build libmbedtls"
+  build_type="Release"
 
-mkdir $libmbedtls_build_dir
-cd $libmbedtls_build_dir
+  mkdir $libmbedtls_build_dir
+  cd $libmbedtls_build_dir
 
-cmake -DCMAKE_BUILD_TYPE=$build_type ${CMAKE_CROSSCOMPILE_OPTION} -DCMAKE_INSTALL_PREFIX=$libmbedtls_install_dir -DENABLE_FLOAT=ON -DBUILD_SHARED_LIBS=OFF ..
-make -j${JOB_NUM} && make install
-if [ $? -ne 0 ];then
-    exit -1
-fi
-
+  cmake -DCMAKE_BUILD_TYPE=$build_type ${CMAKE_CROSSCOMPILE_OPTION} -DCMAKE_INSTALL_PREFIX=$libmbedtls_install_dir -DENABLE_FLOAT=ON -DBUILD_SHARED_LIBS=OFF ..
+  make -j${JOB_NUM} && make install
+  if [ $? -ne 0 ];then
+  	exit -1
+  fi
+fi 
 # libmbedtls end
 
 cd $PROJECT_ROOT_DIR
-build_type="release"
+build_type="debug"
 
 rm -rf build
 meson $MESON_CROSSCOMPILE --buildtype=$build_type build
